test: add failing test for upload_text_to_weaviate_collection
